Move "surrogates" to a dedicated file
